requirements.txt
networkx==2.3
numpy==1.16.4
pandas==0.24.2
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
torch-cluster==1.4.2
torch-scatter==1.3.1
torch-sparse==0.4.0
torch-geometric==1.3.0
tqdm==4.32.2